+++ CHANGELOG.md
@@ -1,7 +1,18 @@
+## 0.2.0 (unreleased)
+
+- Improved accuracy of `Normal::ppf` and `StudentsT::ppf`
+- Return `NAN` instead of panicking for invalid inputs
+
+## 0.1.3 (2022-07-31)
+
+- Added support for `df` between zero and one to `StudentsT::pdf`
+- Fixed error with `StudentsT::cdf` when `x` is infinite or NaN
+- Fixed bug with `StudentsT` functions when `df` is infinity
+
 ## 0.1.2 (2022-07-26)
 
 - Improved accuracy of `Normal::cdf` and `StudentsT::cdf`
-- Fixed bug when `std_dev` is not one
+- Fixed bug with `Normal` functions when `std_dev` is not one
 
 ## 0.1.1 (2022-04-20)